move dl related function into separated file
